% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation)
}
\arguments{
\item{name}{test string, the name of the person to praise.}

\item{punctuation}{emphasis as a text input}
}
\value{
Text string with praise
}
\description{
This is a function to deliver praise, super useful for Nick.
}
\examples{
praise(name = "Steph", punctuation = "!")
}
