% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string, the name of the person to praise.}

\item{punctuation}{text string, emphasis as a text input, default is "!".}
}
\value{
Text string with praise
}
\description{
This is a function to deliver praise, super useful for anyone.
}
\examples{
praise(name = "Steph", punctuation = "!")
}
